code updated in the SC-phase channel, for disorder drivers. Calculation routine for R-DMFT are moved into RDMFT_FUNX module, both for matsubara and real frequency formalism. A RDMFT generic module is created that is to be included to the calling program to use RDMFT layer. The routines in RDMFT_FUNX contain the call to solver interface. Up to now it only use IPT. ED is to come. The drivers are much more simple now, only contain a print routine (plus the search_mu one).
TODO:
The same reshape should be done for trap and REPULSIVE case. This is straighforward, we only need to take care of symmetries to independent sites.
Massimo Capone suggested to implement a check on the lattice symmetries so to avoid specifying the symmetries and making the
codes identical (no more differences between trap and disorder, generic w/ respect to lattice structure).