updated code with latest and working version of the ED_SUPERCONDUCTING CODE
updated structure of the code to fit the introduction of PARSE_INPUT in SciFor.

The input variables from RDMFT and all solver are merged into RDMFT_INPUT_VARS module.

The interface to the specific solver is contained into a wrapping module, e.g. RDMFT_WRAP_IPT,
which in turn call the proper modules to enable use of the solver, e.g. DMFT_IPT.

All the auxiliary functions are store into RDMFT_AUX_FUNX.